Refacto user service : isolate currentuserwithperimeters
